+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## O.0.2
 
+- Command "dataset links" automatically stores links ind tmp/links
 - Travis CI Setup
 
 ## O.0.1